Self code review. Cleaning up comments. Single module async-utils. More test coverage for async-utils.
