+++ test/common/app.ts
@@ -45,11 +45,7 @@ export async function createApp(
   if (useRoutes) {
     // Register error handler
     app.use((err, req, res, next) => {
-      const responseStatus = err.status ?? 500;
-      if (responseStatus === 500) {
-        console.error(err);
-      }
-      res.status(responseStatus).json({
+      res.status(err.status ?? 500).json({
         message: err.message,
         errors: err.errors,
       });
